Pulled fields into seperate component
