+++ index.js
@@ -1,3 +1,4 @@
-import { AppRegistry } from 'react-native';
+import { AppRegistry, YellowBox } from 'react-native';
 import App from './src/App';
+YellowBox.ignoreWarnings(['Warning: isMounted(...) is deprecated', 'Module RCTImageLoader']);
 AppRegistry.registerComponent('manager', () => App);
