+++ src/components/EmployeeListItem.js
@@ -1,19 +1,26 @@
 import React,{Component} from 'react';
-import {Text, View} from 'react-native';
+import {Text, View, TouchableWithoutFeedback} from 'react-native';
+import {Actions} from 'react-native-router-flux';
 import {CardItem} from "./common";
 
 class EmployeeListItem extends Component {
+  onRowPress() {
+      Actions.employeeCreate({ employee: this.props.employee });
+  }
+
   render(){
       const {name} = this.props.employee.item;
 
       return(
-          <View>
-              <CardItem>
-                <Text style={styles.titleStyle} >
-                    {name}
-                </Text>
-              </CardItem>
-          </View>
+          <TouchableWithoutFeedback onPress={this.onRowPress.bind(this)}>
+              <View>
+                  <CardItem>
+                    <Text style={styles.titleStyle} >
+                        {name}
+                    </Text>
+                  </CardItem>
+              </View>
+          </TouchableWithoutFeedback>
       );
   }
 }
